requirements.txt
streamlit==1.31.0
streamlit-authenticator==0.3.1
psycopg2-binary
pandas
langchain
langchain-community
langchain-ollama
langchain-postgres
python-dotenv
passlib
